:boom: Revise interfaces of backup-manager / Add manager-compressor (#84)
* :sparkles: Add manager-backup-compress / :boom: Revise interfaces of backup-manager


* :truck: Rename manager/compressor

* :wheel_of_dharma: Update configmaps


* :truck: migrate cassandra ddl.cql

* :bug: Use hex for manager-cassandra

+++ pkg/manager/backup/cassandra/model/model.go
@@ -18,8 +18,8 @@
 package model
 
 type MetaVector struct {
-	UUID   string    `db:"uuid"`
-	Vector []float64 `db:"vector"`
-	Meta   string    `db:"meta"`
-	IPs    []string  `db:"ips"`
+	UUID   string   `db:"uuid"`
+	Vector []byte   `db:"vector"`
+	Meta   string   `db:"meta"`
+	IPs    []string `db:"ips"`
 }
